2761 Custom Table Component - Turn Customizable column function into HOC - Create CustomTable for custom table functionality
